Improve documentation (#697)
* Add devices doc

* Update doc

* Update doc

* Add more info on net dev

* Use markdown table

* Use ascii table

* Add reserved bit

* Add more content to the device doc

* Update website

* Add console

* Join print args before syscall

* Update syscall functions

* Add structures to syscalls doc

* Update doc

* Add sentence about convertion of struct to ptr and len

* Fix spawn signature

* Add more text

* Use u8 instead of usize for OpenFlag

* Implement BitOr for OpenFlag

* Add info about how to create and read a dir

* Add link to syscalls doc

* Update boot and install

* Add more info on poll syscall

* Fix syscalls doc issues

* Add test for info and kind syscalls

* Add note about return convention

* Fix BMP error messages

* Fix doc wording

* Add documentation for both the raw syscall and its high-level wrapper

* Update doc

* Change dup syscall return type

* Update syscall return types

* Update doc

* Add power module to API for reboot and halt

* Update binaries

* Improve spawn code and doc

* Fix memory count

* Update documentation for network devices

* Use indented code blocks instead of fenced

* Fix DHCP return code

* Fix VGA buffer wording

* Add style to title

* Fix lisp doc

* Update banner

+++ src/usr/http.rs
@@ -134,7 +134,7 @@ pub fn main(args: &[&str]) -> Result<(), ExitCode> {
     };
 
     let mut code = None;
-    let flags = OpenFlag::Device as usize;
+    let flags = OpenFlag::Device as u8;
     if let Some(handle) = syscall::open(socket_path, flags) {
         if syscall::connect(handle, addr, port).is_err() {
             error!("Could not connect to {}:{}", addr, port);
